// app/src/generated/StoredGame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Game } from "./Game";
import type { GameSummary } from "./GameSummary";
import type { Review } from "./Review";

export type StoredGame = { summary: GameSummary, game: Game, review: Review, };